Release: Sprint 4

Closes #98

See merge request informatik/qe/swess23/group4/g4t3!60

+++ src/main/java/at/qe/skeleton/repositories/MeasurementRepository.java
@@ -1,8 +1,8 @@
 package at.qe.skeleton.repositories;
 
-import at.qe.skeleton.model.Measurement;
+import at.qe.skeleton.models.Measurement;
 
 public interface MeasurementRepository extends AbstractRepository<Measurement, Long> {
 
-    Measurement findFirstById(Long id);
+    Measurement findFirstById(Integer id);
 }
